backend/requirements.txt
# AI/ML Libraries
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
onnxruntime==1.16.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
scipy==1.11.4

# NLP & Chatbot
transformers==4.36.0
optimum==1.16.0
sentencepiece==0.1.99

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Sensors & Hardware (for Raspberry Pi)
# RPi.GPIO==0.7.1  # Uncomment when running on Raspberry Pi
# adafruit-circuitpython-max30100==1.1.0
# adafruit-blinka==8.25.0
pyserial==3.5
# spidev==3.6  # Uncomment on Raspberry Pi
# smbus2==0.4.3  # Uncomment on Raspberry Pi

# GPS
# gpsd-py3==0.3.0  # Uncomment on Raspberry Pi
pynmea2==1.19.0

# Communication
requests==2.31.0
aiohttp==3.9.1
python-socketio==5.10.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
reportlab==4.0.7

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code Quality
black==23.12.1
flake8==7.0.0
pylint==3.0.3
